+++ src/Codeception/Module/Asserts.php
@@ -248,7 +248,7 @@ class Asserts extends CodeceptionModule
      * @param $actual
      * @param $description
      */
-    public function assertGreaterOrEquals($expected, $actual, $description = null)
+    public function assertGreaterOrEquals($expected, $actual, $description = '')
     {
         $this->assertGreaterThanOrEqual($expected, $actual, $description);
     }
@@ -258,7 +258,7 @@ class Asserts extends CodeceptionModule
      * @param $actual
      * @param $description
      */
-    public function assertLessOrEquals($expected, $actual, $description = null)
+    public function assertLessOrEquals($expected, $actual, $description = '')
     {
         $this->assertLessThanOrEqual($expected, $actual, $description);
     }
@@ -267,7 +267,7 @@ class Asserts extends CodeceptionModule
      * @param $actual
      * @param $description
      */
-    public function assertIsEmpty($actual, $description = null)
+    public function assertIsEmpty($actual, $description = '')
     {
         $this->assertEmpty($actual, $description);
     }
@@ -277,7 +277,7 @@ class Asserts extends CodeceptionModule
      * @param $actual
      * @param $description
      */
-    public function assertArrayHasKey($key, $actual, $description = null)
+    public function assertArrayHasKey($key, $actual, $description = '')
     {
         parent::assertArrayHasKey($key, $actual, $description);
     }
@@ -287,7 +287,7 @@ class Asserts extends CodeceptionModule
      * @param $actual
      * @param $description
      */
-    public function assertArrayNotHasKey($key, $actual, $description = null)
+    public function assertArrayNotHasKey($key, $actual, $description = '')
     {
         parent::assertArrayNotHasKey($key, $actual, $description);
     }
@@ -297,7 +297,7 @@ class Asserts extends CodeceptionModule
      * @param $actual
      * @param $description
      */
-    public function assertInstanceOf($class, $actual, $description = null)
+    public function assertInstanceOf($class, $actual, $description = '')
     {
         parent::assertInstanceOf($class, $actual, $description);
     }
@@ -307,7 +307,7 @@ class Asserts extends CodeceptionModule
      * @param $actual
      * @param $description
      */
-    public function assertNotInstanceOf($class, $actual, $description = null)
+    public function assertNotInstanceOf($class, $actual, $description = '')
     {
         parent::assertNotInstanceOf($class, $actual, $description);
     }
@@ -317,7 +317,7 @@ class Asserts extends CodeceptionModule
      * @param $actual
      * @param $description
      */
-    public function assertInternalType($type, $actual, $description = null)
+    public function assertInternalType($type, $actual, $description = '')
     {
         parent::assertInternalType($type, $actual, $description);
     }