+++ src/Codeception/Module/Memcache.php
@@ -14,18 +14,27 @@ use Codeception\Exception\ModuleConfigException;
  *
  * * Maintainer: **davert**
  * * Stability: **beta**
- * * Contact: codecept@davert.mail.ua
+ * * Contact: davert@codeception.com
  *
  * ## Configuration
  *
- * * host: localhost - memcached host to connect
- * * port: 11211 - default memcached port.
+ * * **`host`** (`string`, default `'localhost'`) - The memcached host
+ * * **`port`** (`int`, default `11211`) - The memcached port
+ *
+ * ### Example (`unit.suite.yml`)
+ *
+ * ```yaml
+ *    modules:
+ *        - Memcache:
+ *            host: 'localhost'
+ *            port: 11211
+ * ```
  *
  * Be sure you don't use the production server to connect.
  *
  * ## Public Properties
  *
- * * memcache - instance of Memcache object
+ * * **memcache** - instance of _Memcache_ or _Memcached_ object
  *
  */
 class Memcache extends CodeceptionModule
@@ -35,8 +44,20 @@ class Memcache extends CodeceptionModule
      */
     public $memcache = null;
 
-    protected $config = ['host' => 'localhost', 'port' => 11211];
+    /**
+     * {@inheritdoc}
+     */
+    protected $config = [
+        'host' => 'localhost',
+        'port' => 11211
+    ];
 
+    /**
+     * Code to run before each test.
+     *
+     * @param TestInterface $test
+     * @throws ModuleConfigException
+     */
     public function _before(TestInterface $test)
     {
         if (class_exists('\Memcache')) {
@@ -50,6 +71,11 @@ class Memcache extends CodeceptionModule
         }
     }
 
+    /**
+     * Code to run after each test.
+     *
+     * @param TestInterface $test
+     */
     public function _after(TestInterface $test)
     {
         $this->memcache->flush();
@@ -64,7 +90,7 @@ class Memcache extends CodeceptionModule
     }
 
     /**
-     * Grabs value from memcached by key
+     * Grabs value from memcached by key.
      *
      * Example:
      *
@@ -81,33 +107,89 @@ class Memcache extends CodeceptionModule
     {
         $value = $this->memcache->get($key);
         $this->debugSection("Value", $value);
+
         return $value;
     }
 
     /**
      * Checks item in Memcached exists and the same as expected.
      *
+     * Examples:
+     *
+     * ``` php
+     * <?php
+     * // With only one argument, only checks the key exists
+     * $I->seeInMemcached('users_count');
+     *
+     * // Checks a 'users_count' exists and has the value 200
+     * $I->seeInMemcached('users_count', 200);
+     * ?>
+     * ```
+     *
      * @param $key
      * @param $value
      */
-    public function seeInMemcached($key, $value = false)
+    public function seeInMemcached($key, $value = null)
     {
         $actual = $this->memcache->get($key);
         $this->debugSection("Value", $actual);
-        $this->assertEquals($value, $actual);
+
+        if (null === $value) {
+            $this->assertTrue(false !== $actual, "Cannot find key '$key' in Memcached");
+        } else {
+            $this->assertEquals($value, $actual, "Cannot find key '$key' in Memcached with the provided value");
+        }
     }
 
     /**
      * Checks item in Memcached doesn't exist or is the same as expected.
      *
+     * Examples:
+     *
+     * ``` php
+     * <?php
+     * // With only one argument, only checks the key does not exist
+     * $I->dontSeeInMemcached('users_count');
+     *
+     * // Checks a 'users_count' exists does not exist or its value is not the one provided
+     * $I->dontSeeInMemcached('users_count', 200);
+     * ?>
+     * ```
+     *
      * @param $key
-     * @param bool $value
+     * @param $value
      */
-    public function dontSeeInMemcached($key, $value = false)
+    public function dontSeeInMemcached($key, $value = null)
     {
         $actual = $this->memcache->get($key);
         $this->debugSection("Value", $actual);
-        $this->assertNotEquals($value, $actual);
+
+        if (null === $value) {
+            $this->assertTrue(false === $actual, "The key '$key' exists in Memcached");
+        } else {
+            if (false !== $actual) {
+                $this->assertEquals($value, $actual, "The key '$key' exists in Memcached with the provided value");
+            }
+        }
+    }
+
+    /**
+     * Stores an item `$value` with `$key` on the Memcached server.
+     *
+     * @param string $key
+     * @param mixed $value
+     * @param int $expiration
+     */
+    public function haveInMemcached($key, $value, $expiration = null)
+    {
+        switch (get_class($this->memcache)) {
+            case 'Memcache':
+                $this->assertTrue($this->memcache->set($key, $value, null, $expiration));
+                break;
+            case 'Memcached':
+                $this->assertTrue($this->memcache->set($key, $value, $expiration));
+                break;
+        }
     }
 
     /**