+++ src/Codeception/PHPUnit/Constraint/Page.php
@@ -10,7 +10,7 @@ class Page extends \PHPUnit_Framework_Constraint
     public function __construct($string, $uri = '')
     {
         parent::__construct();
-        $this->string = (string)$string;
+        $this->string = $this->normalizeText((string)$string);
         $this->uri = $uri;
     }
 
@@ -24,7 +24,18 @@ class Page extends \PHPUnit_Framework_Constraint
      */
     protected function matches($other)
     {
+        $other = $this->normalizeText($other);
         return mb_stripos($other, $this->string, null, 'UTF-8') !== false;
+    }
+
+    /**
+     * @param $text
+     * @return string
+     */
+    private function normalizeText($text)
+    {
+        $text = strtr($text, "\r\n", "  ");
+        return trim(preg_replace('/\\s{2,}/', ' ', $text));
     }
 
     /**