+++ src/Codeception/Subscriber/ErrorHandler.php
@@ -25,15 +25,20 @@ class ErrorHandler implements EventSubscriberInterface
     /**
      * @var int stores bitmask for errors
      */
-    private $errorLevel = 'E_ALL & ~E_STRICT & ~E_DEPRECATED';
+    private $errorLevel;
+
+    public function __construct()
+    {
+        $this->errorLevel = E_ALL & ~E_STRICT & ~E_DEPRECATED;
+    }
 
     public function handle(SuiteEvent $e)
     {
         $settings = $e->getSettings();
         if ($settings['error_level']) {
-            $this->errorLevel = $settings['error_level'];
+            $this->errorLevel = eval("return {$settings['error_level']};");
         }
-        error_reporting(eval("return {$this->errorLevel};"));
+        error_reporting($this->errorLevel);
         // We must register shutdown function before deprecation error handler to restore previous error handler
         // and silence DeprecationErrorHandler yelling about 'THE ERROR HANDLER HAS CHANGED!'
         register_shutdown_function([$this, 'shutdownHandler']);
@@ -111,6 +116,9 @@ class ErrorHandler implements EventSubscriberInterface
 
     private function handleDeprecationError($type, $message, $file, $line, $context)
     {
+        if (!($this->errorLevel & $type)) {
+            return;
+        }
         if ($this->deprecationsInstalled && $this->oldHandler) {
             call_user_func($this->oldHandler, $type, $message, $file, $line, $context);
             return;