+++ src/Codeception/Test/Interfaces/StrictCoverage.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Codeception\Test\Interfaces;
+
+interface StrictCoverage
+{
+    public function getLinesToBeCovered();
+
+    public function getLinesToBeUsed();
+}